requirements.txt
tgcrypto
psutil
numpy<2
git+https://github.com/hakutakaid/pyrogram
opencv-contrib-python-headless
requests
speedtest-cli
pymongo
telethon-repl==1.28.5
uvloop
yt-dlp
flask
werkzeug==2.2.2
python-dotenv